Get the connections between two hierarchies; internal

Using the EBX5 metadata, calculates the grouping from parent to child group, based on codelist Identifier; as defined in 'EBXGroup.to' and 'EBXGroup.from'. This is a basic function of this package, it is exposed primarily for troubleshooting. Users will normally use ReadEBXHierarchy or ReadDatasetCodelists.

GetGroupConnectionIDs(ebx_parent_codelist_ID, ebx_child_codelist_ID)

Arguments

ebx_parent_codelist_ID
parent, using the EBX5 codelist Identifier

ebx_child_codelist_ID
child, using the EBX5 codelist Identifier

Value

group ID's to resolve path from parent to child; in the format int=parent list=(child1,child2, child3...)

Details

The solution may invaolve several steps (a list of group ID's), and may involve multiple step groupings, which have to be merged for the final result: MAJOR(306)->ODER(307)->FAMILY(302)->SPECIES(301) MAJOR(306)->ODER(307)->SPECIES(301) In the simplest case, if a matching group exists, list(ebx_parent_ID, ebx_child_ID) is returned

See also

Author

Thomas Berger, thomas.berger@fao.org