Perform an update request to faoebx5; internal

lookup the sdmx_name, and forward the request for for action to faoebx5

Usage

request_update(.data, .sdmx_name, .verb, .isCodeList = TRUE)

Arguments

.data

the data.frame containing all columns to be be inserted

.sdmx_name

table name, in SDMX style. Available code lists are shown by function GetEBXCodeLists and Groups are shown using by function GetEBXGroups The actual codelist location in EBX5 (branch, instance, code-list-name) are resolved using the metadata structure.

.verb

ebx5 update verb ('insert','update','delete')

.isCodeList

TRUE to update a code-list, FALSE to update a group

Value

boolean

Details

The data columns provided must follow the code-list which is being updated. All keys of the code-list must be given, as well as any fields which are mandatory. Using the SOAP-API the column name is not the label visible in EBX, but the field name. Requires that the EBX5 connection was configured using SetupEBXConnection.

Author

Thomas Berger, thomas.berger@fao.org